pyproject.toml
[project]
name = "saturn-tts"
version = "0.1.0"
description = "Emotional text-to-speech via style-conditioned flow matching"
readme = "README.md"
license = "CC-BY-NC-4.0"
requires-python = ">=3.10"
authors = [
    { name = "Jose Alarcon Silva" },
]

dependencies = [
    "torch>=2.0",
    "torchaudio",
    "f5-tts",
    "vocos",
    "safetensors",
    "ema-pytorch",
    "x-transformers>=1.31",
    "torchdiffeq",
    "librosa",
    "soundfile",
    "einops",
    "pyyaml",
]

[project.optional-dependencies]
chat = [
    "gradio>=6.0",
    "transformers>=5.5",
    "openai-whisper",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.setuptools.packages.find]
where = ["src"]
